### 2023-05-03 #5431 @jeffhammond
Himscipy, Chilipp, gonsie, csadorf, kris-rowe all list Python as preferred language and show expertise in MPI somewhere.  They are listed in the order they appear on the sheet.
