### 2023-05-04 #5444 @editorialbot
[REVIEW]: shmem4py: OpenSHMEM for Python
**Submitting author:** @mrogowski (<a href="http://orcid.org/0000-0002-5662-2082">Marcin Rogowski</a>)
**Repository:** https://github.com/mpi4py/shmem4py
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.0.0
