+++ .github/workflows/native.yml
@@ -8,6 +8,8 @@ on:
     branches:
       - master
   workflow_dispatch:
+  schedule:
+    - cron: '33 3 * * 0'
 
 jobs:
 
