### 2023-06-20 #5444 @danielskatz
👋 @greghbauer - Can you generate your checklist, and ideally check off the first two items, just to make sure all the permissions here are working right?
