### 2023-06-27 #5444 @gonsie
@mrogowski

Seems like the **Community Guidelines** stuff is missing (I couldn't find any guidance on the documentation site, nor in the repo itself).
