### 2023-06-28 #5444 @greghbauer
After several false starts, I was able to complete the functionality testing, using Open MPI OpenSHMEM on NCSA Delta with anaconda, with SLURM srun replacing oshrun. I was able to do some of the tests across more than one node.

The install instructions at https://shmem4py.readthedocs.io/en/latest/installation.html state "Once a working OpenSHMEM implementation is installed" trivializes the task.

The install instructions at https://github.com/mpi4py/shmem4py/blob/master/INSTALL.rst are more complete but assume root level capabilities with the user of **sudo apt-get**.

I did not try to use the docker container docker build scripts. On HPC sites docker is typically not possible. Would sif build scripts be possible?
