### 2023-07-02 eaf2df8
Add CONTRIBUTING.md (#17)
Add CONTRIBUTING.md

Co-authored-by: Lisandro Dalcin <dalcinl@gmail.com>
