+++ CONTRIBUTING.md
@@ -13,9 +13,9 @@ Thanks for taking the time to contribute!
 
 > If you want to ask a question, we assume you have read the available [documentation](https://shmem4py.readthedocs.io/).
 
-Before you ask a question, it is best to search for existing [Issues](/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first. Please note that many questions regarding the programming model itself and suggestions for its enhancements may better belong to the [OpenSHMEM Community](http://openshmem.org/site/Community) rather than the shmem4py project, which is naturally constrained by the specification and underlying OpenSHMEM implementations.
+Before you ask a question, it is best to search for existing [Issues](https://github.com/mpi4py/shmem4py/issues) that might help you. In case you have found a suitable issue and still need clarification, you can write your question in this issue. It is also advisable to search the internet for answers first. Please note that many questions regarding the programming model itself and suggestions for its enhancements may better belong to the [OpenSHMEM Community](http://openshmem.org/site/Community) rather than the shmem4py project, which is naturally constrained by the specification and underlying OpenSHMEM implementations.
 
-If you still feel the need to ask a question or clarification, we recommend opening a [Discussion (Q&A)](/discussions/new?category=q-a). Please remember to provide as much context as possible about what you're running into and report project and platform versions depending on what seems relevant.
+If you still feel the need to ask a question or clarification, we recommend opening a [Discussion (Q&A)](https://github.com/mpi4py/shmem4py/discussions/new?category=q-a). Please remember to provide as much context as possible about what you're running into and report project and platform versions depending on what seems relevant.
 
 We will then take care of the issue as soon as possible.
 
@@ -32,7 +32,7 @@ A good bug report shouldn't leave others needing to chase you for more informati
 
 - Make sure that you are using the latest version.
 - Determine if your bug is really a bug and not an error on your side, e.g., using incompatible environment components/versions (Make sure that you have read the [documentation](https://shmem4py.readthedocs.io/).
-- To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](issues?q=label%3Abug).
+- To see if other users have experienced (and potentially already solved) the same issue you are having, check if there is not already a bug report existing for your bug or error in the [bug tracker](https://github.com/mpi4py/shmem4py/issues).
 - Also, make sure to search the internet (including Stack Overflow) to see if users outside of the GitHub community have discussed the issue.
 - Collect information about the bug:
   - Stack trace (Traceback)
@@ -46,7 +46,7 @@ A good bug report shouldn't leave others needing to chase you for more informati
 
 We use GitHub issues to track bugs and errors. If you run into an issue with the project:
 
-- Open an [Issue](/issues/new).
+- Open an [Issue](https://github.com/mpi4py/shmem4py/issues/new).
 - Explain the behavior you would expect and the actual behavior.
 - Please provide as much context as possible and describe the *reproduction steps* that someone else can follow to recreate the issue on their own. This usually includes your code. For good bug reports, you should isolate the problem and create a reduced test case.
 - Provide the information you collected in the previous section.
@@ -60,14 +60,14 @@ This section guides you through submitting an enhancement suggestion for shmem4p
 
 - Make sure that you are using the latest version.
 - Read the [documentation](https://shmem4py.readthedocs.io/) carefully and find out if the functionality is already covered, maybe by an individual configuration.
-- Perform a [search](/issues) to see if the enhancement has already been suggested. If it has, add a comment to the existing issue instead of opening a new one.
[…]
