### 2023-07-02 #5444 @mrogowski
We have added a [CONTRIBUTING.md](https://github.com/mpi4py/shmem4py/blob/master/CONTRIBUTING.md) file in the root directory of the repository. We hope it clarifies how to seek support, report issues and contribute to the software. We used a template recommended by other JOSS reviewers. Thanks for pointing it out, @gonsie @greghbauer!
