### 2023-07-05 #5444 @greghbauer
Hi.
I am ok with the changes.
Greg

On Mon, Jul 3, 2023 at 10:31 AM Daniel S. Katz ***@***.***>
wrote:

> @gonsie <https://github.com/gonsie> & @greghbauer
> <https://github.com/greghbauer> - is the contributing part now ok?
>
> @gonsie <https://github.com/gonsie> - I note you haven't checked off any
> of the functionality part of your list - Is something blocking this part
> for you?
>
> @greghbauer <https://github.com/greghbauer> - Re your comments above
> <https://github.com/openjournals/joss-reviews/issues/5444#issuecomment-1610628425>,
> are any of these issues that you think should block acceptance of this
> work? Or are they more suggestions for the future?
>
