### 2023-07-06 #5444 @danielskatz
👋 @gonsie - is the contributing part now ok?

And I note you haven't checked off any of the functionality part of your list - Is something blocking this part for you?
