### 2023-07-10 #5444 @mrogowski
@greghbauer thanks for approving all the items and the latest changes! I do not intend to ignore your comment on installation instructions. I will modify the text assuming configured MPI (clusters) or root access (workstations). I think those are fair assumptions unlike "a working OpenSHMEM implementation". 🙂
