### 2023-07-10 #5444 @gonsie
I'm running on a Mac... no experience with Docker. I also have access to a RedHat x86_64 cluster, but no root access.
