### 2023-07-10 #5444 @mrogowski
Could you please try to follow the steps below on your Mac? I just set it up on my machine and it works fine, even though I have not tested on a Mac before. The first step installs a few packages with [brew](https://brew.sh/).

```
  brew install gcc autoconf automake libtool mpich
  git clone https://github.com/pmodels/oshmpi --recurse-submodules
  cd oshmpi
  ./autogen.sh
  mkdir build && cd build
  ../configure CC=/opt/homebrew/bin/mpicc CXX=/opt/homebrew/bin/mpicxx --prefix=/Users/XXXXXXXX/SHMEM/install-oshmpi
  make -j
  make install
  export PATH=/Users/XXXXXXXX/SHMEM/install-oshmpi/bin/:$PATH
  export FI_PROVIDER=tcp
  git clone https://github.com/mpi4py/shmem4py/
  cd shmem4py
  python -m pip install .
  cd demo
  mpiexec -n 3 python hello.py
  mpiexec -n 8 python race_winner.py
```
