### 2023-07-10 mpi4py/shmem4py#19 @gonsie
Building on a Mac
I'm moving the conversation from openjournals/joss-reviews#5444 to this thread... which may be useful to others in the future.

I'm using [these instructions](https://github.com/openjournals/joss-reviews/issues/5444#issuecomment-1629407483) from @mrogowski:

```
  brew install gcc autoconf automake libtool mpich
  git clone https://github.com/pmodels/oshmpi --recurse-submodules
  cd oshmpi
  ./autogen.sh
  mkdir build && cd build
  ../configure CC=/opt/homebrew/bin/mpicc CXX=/opt/homebrew/bin/mpicxx --prefix=/Users/XXXXXXXX/SHMEM/install-oshmpi
  make -j
  make install
  export PATH=/Users/XXXXXXXX/SHMEM/install-oshmpi/bin/:$PATH
  export FI_PROVIDER=tcp
  git clone https://github.com/mpi4py/shmem4py/
  cd shmem4py
  python -m pip install .
  cd demo
  mpiexec -n 3 python hello.py
  mpiexec -n 8 python race_winner.py
```
