### 2023-07-10 mpi4py/shmem4py#19 @gonsie
First error, see #18

Now I'm getting the following:
```
fatal error: 'shmem.h' file not found
      #include <shmem.h>
```

Note that I'm using clang instead of GCC
