### 2023-07-10 f6cc4e4
Update setup.cfg for README.rst (#18)
README file type has changed
