+++ setup.cfg
@@ -2,7 +2,7 @@
 name = shmem4py
 version = attr: shmem4py.__version__
 description = Python bindings for OpenSHMEM
-long_description = file: README.md
+long_description = file: README.rst
 long_description_content_type = text/markdown
 license = BSD-2-Clause
 license_files = LICENSE.md
