### 2023-07-10 mpi4py/shmem4py#19 @mrogowski
I suspect that something went wrong with your `PATH`. Is `oshcc` in the PATH?
