### 2023-07-10 mpi4py/shmem4py#19 @mrogowski
I had similar issues due to version mismatches between MPICH and libfabric... Can you try to set those before running?
```
export MPIR_CVAR_OFI_SKIP_IPV6=0
export FI_PROVIDER=tcp
```
