### 2023-07-10 mpi4py/shmem4py#19 @gonsie
Seems like it's an issue with my local MPI install
