### 2023-07-10 mpi4py/shmem4py#19 @mrogowski
Another [solution](https://lists.mpich.org/pipermail/discuss/2021-May/006198.html) suggests `export MPIR_CVAR_OFI_USE_PROVIDER=sockets` and/or `export FI_PROVIDER=sockets`.
