### 2023-07-10 mpi4py/shmem4py#19 @gonsie
The `export FI_PROVIDER=tcp` was breaking things. now the demos are working! 🎉
