### 2023-07-10 mpi4py/shmem4py#19 @gonsie
For posterity, these are the steps that worked for me:

```
# mpich previously installed without fortran, thus negating the need to install gcc
 brew install autoconf automake libtool
  git clone https://github.com/pmodels/oshmpi --recurse-submodules
  cd oshmpi
  ./autogen.sh
  mkdir install
  mkdir build && cd build
  ../configure CC=/opt/homebrew/bin/mpicc CXX=/opt/homebrew/bin/mpicxx --prefix=/Users/XXXXXXXX/oshmpi/install/
  make -j
  make install
  export PATH=/Users/XXXXXXXX/oshmpi/install/bin/:$PATH
  git clone https://github.com/mpi4py/shmem4py/
  cd shmem4py
# create a virtual environment if you prefer
  python -m pip install .
  cd demo
  mpiexec -n 3 python hello.py
  mpiexec -n 8 python race_winner.py
```
