### 2023-07-11 #5444 @mrogowski
@danielskatz, thanks for persistently following up with all of us! It looks like all the items on both checklists are checked off now. I am yet to modify the installation instructions to reflect @greghbauer's comments and @gonsie's experience on a Mac. What other steps should we plan for regarding the release and publication?
