### 2023-07-12 #5444 @gonsie
Yes, I think the Mac install instructions would be helpful, but I'm happy to accept this. Thank you.
