### 2023-07-13 8cf1963
Update setup.cfg (#20)
