+++ setup.cfg
@@ -3,7 +3,7 @@ name = shmem4py
 version = attr: shmem4py.__version__
 description = Python bindings for OpenSHMEM
 long_description = file: README.rst
-long_description_content_type = text/markdown
+long_description_content_type = text/x-rst
 license = BSD-2-Clause
 license_files = LICENSE.md
 url = https://github.comm/mpi4py/shmem4py
