+++ docker/oshmem_ubuntu/Dockerfile
@@ -13,8 +13,9 @@ RUN apt-get update -y && apt-get install -y \
   python3 python3-pip python-is-python3
 
 RUN cd $INSTALL_DIR                                                                                                  && \
-    git clone https://github.com/openucx/ucx                                                                         && \
-    cd ucx && ./autogen.sh                                                                                           && \
+    wget https://github.com/openucx/ucx/archive/refs/tags/v1.14.1.tar.gz                                             && \
+    tar xf v1.14.1.tar.gz                                                                                            && \
+    cd ucx-1.14.1 && ./autogen.sh                                                                                    && \
     ./configure --prefix=$INSTALL_DIR/ucx-master/install --disable-debug --disable-assertions --disable-params-check && \
     make -j && make install
 
