+++ docker/oshmem_ubuntu/Dockerfile
@@ -12,11 +12,11 @@ RUN apt-get update -y && apt-get install -y \
   wget                                      \
   python3 python3-pip python-is-python3
 
-RUN cd $INSTALL_DIR                                                                                           && \
-    wget https://github.com/openucx/ucx/archive/refs/tags/v1.14.1.tar.gz                                      && \
-    tar xf v1.14.1.tar.gz                                                                                     && \
-    cd ucx-1.14.1 && ./autogen.sh                                                                             && \
-    ./configure --prefix=$INSTALL_DIR/ucx/install --disable-debug --disable-assertions --disable-params-check && \
+RUN cd $INSTALL_DIR                                                                                                                         && \
+    wget https://github.com/openucx/ucx/archive/refs/tags/v1.14.1.tar.gz                                                                    && \
+    tar xf v1.14.1.tar.gz                                                                                                                   && \
+    cd ucx-1.14.1 && ./autogen.sh                                                                                                           && \
+    ./configure CFLAGS=-Wno-enum-int-mismatch --prefix=$INSTALL_DIR/ucx/install --disable-debug --disable-assertions --disable-params-check && \
     make -j && make install
 
 RUN cd $INSTALL_DIR                                                                   && \
