+++ docker/oshmem_fedora/Dockerfile
@@ -8,11 +8,11 @@ RUN dnf update -y && \
                    pmix pmix-devel libevent libevent-devel python3 python3-devel python3-pip && \
     dnf clean all
 
-RUN cd $INSTALL_DIR                                                                                                                         && \
-    wget https://github.com/openucx/ucx/archive/refs/tags/v1.14.1.tar.gz                                                                    && \
-    tar xf v1.14.1.tar.gz                                                                                                                   && \
-    cd ucx-1.14.1 && ./autogen.sh                                                                                                           && \
-    ./configure CFLAGS=-Wno-enum-int-mismatch --prefix=$INSTALL_DIR/ucx/install --disable-debug --disable-assertions --disable-params-check && \
+RUN cd $INSTALL_DIR                                                                                                             && \
+    wget https://github.com/openucx/ucx/archive/refs/tags/v1.14.1.tar.gz                                                        && \
+    tar xf v1.14.1.tar.gz                                                                                                       && \
+    cd ucx-1.14.1 && ./autogen.sh                                                                                               && \
+    ./configure CFLAGS=-Wno-error --prefix=$INSTALL_DIR/ucx/install --disable-debug --disable-assertions --disable-params-check && \
     make -j && make install
 
 RUN cd $INSTALL_DIR                                                                   && \
