+++ demo/lock.py
@@ -1,18 +1,14 @@
 # adapted from http://openshmem.org/site/sites/default/site_files/OpenSHMEM-1.5.pdf (Example 45)
 
-import numpy as np
 from shmem4py import shmem
 
 lock = shmem.new_lock()
 mype = shmem.my_pe()
 
-val = np.array([0], dtype='i')
 count = shmem.array([0], dtype='i')
-shmem.barrier_all()
-print(f"{mype}: past the barrier")
+val = shmem.array([0], dtype='i')
 
 shmem.set_lock(lock)
-print(f"{mype}: set the lock")
 shmem.get(val, count, 0)
 print(f"{mype}: count is {val[0]}")
 val[0] += 1
@@ -21,3 +17,4 @@ shmem.clear_lock(lock)
 
 shmem.del_lock(lock)
 shmem.free(count)
+shmem.free(val)
