+++ .github/workflows/containers.yml
@@ -24,9 +24,9 @@ jobs:
         uses: actions/checkout@v3
       - uses: jpribyl/action-docker-layer-caching@v0.1.1
         with:
-          key: ${{ matrix.shmem }}-4-{hash}
+          key: ${{ matrix.shmem }}-5-{hash}
           restore-keys: |
-            ${{ matrix.shmem }}-4-
+            ${{ matrix.shmem }}-5-
       - name: Build Docker container
         run: docker build -t local docker/${{ matrix.shmem }}/
       - name: run shmem4py test-1
