### 2023-07-13 #5444 @mrogowski
We have published a release with a [1.0.0 tag](https://github.com/mpi4py/shmem4py/releases/tag/1.0.0). It is archived in [Zenodo](https://zenodo.org/record/8143862), and the DOI is [10.5281/zenodo.8143862](https://doi.org/10.5281/zenodo.8143862).
