### 2023-07-13 e0d6bce
Correct the version to 1.0.0
