### 2023-07-17 #5444 @danielskatz
I will proofread this, and then talk about next steps
