### 2023-07-17 #5444 @danielskatz
It's not clear to me where the DOI for this came from, as the original from USENIX (https://www.usenix.org/conference/osdi18/presentation/alquraan) doesn't list a DOI. If this isn't resolve fairly quickly, the best option would probably be to remove the DOI from the bibtex entry and to add this URL instead, but let's see what USENIX says first.
