### 2023-07-19 #5444 @danielskatz
@mrogowski - Let's go ahead and proceed as above. Can you remove the DOI that doesn't work and add the URL?
