### 2023-07-19 c802430
update ray bibtex entry
