+++ joss/paper.bib
@@ -18,19 +18,17 @@
     doi={10.25080/Majora-7b98e3ed-013}
 }
 
-@inproceedings{ray,
-author = {Moritz, Philipp and Nishihara, Robert and Wang, Stephanie and Tumanov, Alexey and Liaw, Richard and Liang, Eric and Elibol, Melih and Yang, Zongheng and Paul, William and Jordan, Michael I. and Stoica, Ion},
-title = {Ray: A Distributed Framework for Emerging AI Applications},
+@inproceedings {ray,
+author = {Philipp Moritz and Robert Nishihara and Stephanie Wang and Alexey Tumanov and Richard Liaw and Eric Liang and Melih Elibol and Zongheng Yang and William Paul and Michael I. Jordan and Ion Stoica},
+title = {Ray: A Distributed Framework for Emerging {AI} Applications},
+booktitle = {13th USENIX Symposium on Operating Systems Design and Implementation (OSDI 18)},
 year = {2018},
-isbn = {9781931971478},
+isbn = {978-1-939133-08-3},
+address = {Carlsbad, CA},
+pages = {561--577},
+url = {https://www.usenix.org/conference/osdi18/presentation/moritz},
 publisher = {USENIX Association},
-address = {USA},
-booktitle = {Proceedings of the 13th USENIX Conference on Operating Systems Design and Implementation},
-pages = {561–577},
-numpages = {17},
-location = {Carlsbad, CA, USA},
-series = {OSDI'18},
-doi = {10.5555/3291168.3291210}
+month = oct,
 }
 
 @article{mpi4py,
