### 2023-07-19 #5444 @mrogowski
I just updated the bibtex entry to that listed on https://www.usenix.org/conference/osdi18/presentation/moritz, including the URL. Thanks for following up on that.
