### 2023-07-19 #5444 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
